+++ paper/src/main/java/io/github/idankoblik/jukebox/PaperSong.java
@@ -22,8 +22,8 @@ public class PaperSong extends KyoriSong<PaperPlatform> {
      * @param plugin the plugin that handles the song
      * {@inheritDoc}
      */
-    public PaperSong(@NotNull Plugin plugin, float volume, @NotNull NBSSequenceWrapper wrapper, @NotNull Key defaultSound, @NotNull Audience audience, @Nullable Position position) {
-        super(wrapper, volume, defaultSound, audience, position);
+    public PaperSong(@NotNull Plugin plugin, float volume, @NotNull NBSSequencePlayer player, @NotNull Key defaultSound, @NotNull Audience audience, @Nullable Position position) {
+        super(player, volume, defaultSound, audience, position);
         if (volume < 0.0f || volume > 1.0f)
             throw new IllegalArgumentException("Volume must be between 0.0 and 1.0");
 
@@ -34,12 +34,12 @@ public class PaperSong extends KyoriSong<PaperPlatform> {
      * {@inheritDoc}
      */
     @Override
-    public CompletableFuture<NBSSequenceWrapper> handle(float volume) {
+    public CompletableFuture<NBSSequencePlayer> handle(float volume) {
         List<NBSNote> notes = sequence.getNotes();
 
-        CompletableFuture<NBSSequenceWrapper> future = new CompletableFuture<>();
+        CompletableFuture<NBSSequencePlayer> future = new CompletableFuture<>();
         if (notes.isEmpty()) {
-            future.complete(wrapper);
+            future.complete(player);
             return future;
         }