+++ paper/src/main/java/io/github/idankoblik/jukebox/SongTick.java
@@ -1,6 +1,5 @@
 package io.github.idankoblik.jukebox;
 
-import io.github.idankoblik.jukebox.events.EventManager;
 import io.github.idankoblik.jukebox.events.SongEndEvent;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.jetbrains.annotations.NotNull;
@@ -14,7 +13,7 @@ import java.util.concurrent.CompletableFuture;
 /* package-private */ class SongTick extends BukkitRunnable {
     private final List<NBSNote> notes;
     private final PaperSong abstractSong;
-    private final CompletableFuture<NBSSequenceWrapper> completionSignal;
+    private final CompletableFuture<NBSSequencePlayer> completionSignal;
     private int tick;
     private final float tickLengthInSeconds;
 
@@ -24,7 +23,7 @@ import java.util.concurrent.CompletableFuture;
      * @param abstractSong the song to be handled
      * @param completionSignal a future of the song
      */
-    public SongTick(@NotNull List<NBSNote> notes, PaperSong abstractSong, CompletableFuture<NBSSequenceWrapper> completionSignal) {
+    public SongTick(@NotNull List<NBSNote> notes, PaperSong abstractSong, CompletableFuture<NBSSequencePlayer> completionSignal) {
         this.notes = notes;
         this.abstractSong = abstractSong;
         this.completionSignal = completionSignal;
@@ -37,8 +36,8 @@ import java.util.concurrent.CompletableFuture;
      */
     @Override
     public void run() {
-        NBSSequenceWrapper wrapper = abstractSong.wrapper;
-        if (wrapper.getState() == SongState.ENDED) {
+        NBSSequencePlayer player = abstractSong.player;
+        if (player.getState() == SongState.ENDED) {
             completeSong();
             return;
         }
@@ -54,8 +53,8 @@ import java.util.concurrent.CompletableFuture;
 
         if (tick > notes.stream().mapToLong(note -> Math.round(note.getTick() * tickLengthInSeconds)).max().orElse(0)) {
             if (!abstractSong.toLoop()) {
-                wrapper.getEventManager().fireEvent(new SongEndEvent(abstractSong.sequence, false));
-                wrapper.setState(SongState.ENDED);
+                player.getEventManager().fireEvent(new SongEndEvent(abstractSong.sequence, false));
+                player.setState(SongState.ENDED);
                 completeSong();
             } else
                 tick = 0;
@@ -77,7 +76,7 @@ import java.util.concurrent.CompletableFuture;
      * Completing the song
      */
     private void completeSong() {
-        completionSignal.complete(abstractSong.wrapper);
+        completionSignal.complete(abstractSong.player);
         cancel();
     }
 }