+++ paper/src/test/java/io/github/idankoblik/jukebox/SongTest.java
@@ -28,18 +28,18 @@ public class SongTest extends AbstractPaperSongTest {
     }
 
     @Test
-    protected void testPlaySong() {
-        super.testPlaySong();
-    }
-
-    @Test
     protected void testTriggeringEvents() {
         super.testTriggeringEvents();
     }
 
+    @Test
+    protected void testPlaySong() {
+        super.testPlaySong();
+    }
+
     @Override
     protected void initializeSongs() {
-        this.song = new PaperSong(plugin, 1F, this.wrapper, Key.key(DEFAULT_NOTE), player, null);
-        this.locationSong = new PaperSong(plugin, 1, this.wrapper, Key.key(DEFAULT_NOTE), world, musicPosition);
+        this.song = new PaperSong(plugin, 1F, this.sequencePlayer, Key.key(DEFAULT_NOTE), player, null);
+        this.locationSong = new PaperSong(plugin, 1, this.sequencePlayer, Key.key(DEFAULT_NOTE), world, musicPosition);
     }
 }